requirements.txt
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.18
openpyxl>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.0
notion-client>=1.0.0
google-api-python-client>=2.0.0
atlassian-python-api>=3.0.0